[core] move pre-parsers into their own file and rename class
* this change reduces complexity in classifier
* uses a consistent naming scheme for pre-parser classes/methods
* updates unit tests to reflect changes
